add atomic validation to daily payout query
When using the command, last_payout_at is validated to make sure the
user is eligible for a payout. However, it is possible for two command
executions to pass the validation check in a race condition, leading to
the user effectively registering two daily payouts at once. The
validation is now done atomically at the query level as well to prevent
this scenario.